refactor(api): adjust api types
each module has its own types.ts

resolve #66
