+++ .husky/pre-commit
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-yarn check --integrity && yarn lint
+yarn check --integrity && yarn pretty-quick --staged && yarn lint
