--- .prettierrc.yml
@@ -1,8 +1,0 @@
-bracketSpacing: true
-trailingComma: 'es5'
-tabWidth: 2
-tabs: false
-semi: true
-singleQuote: true
-jsxSingleQuote: false
-jsxBracketSameLine: true
