--- commitlint.config.js
@@ -1,1 +1,0 @@
-module.exports = { extends: ['@commitlint/config-conventional'] };
