+++ package.json
@@ -85,11 +85,12 @@
     "husky": "^6.0.0",
     "nodemon": "^2.0.7",
     "prettier": "^2.3.1",
+    "pretty-quick": "^3.1.1",
     "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.0",
     "react-custom-scrollbars": "^4.2.1",
     "react-dom": "^17.0.2",
-    "react-intl": "^5.20.3",
+    "react-intl": "^5.20.4",
     "react-markdown": "^6.0.2",
     "react-progressive-image": "^0.6.0",
     "react-redux": "^7.2.4",
