+++ src/api/unsplash/types.ts
@@ -1,0 +1,14 @@
+export interface IUnsplashPhoto {
+  id: string;
+  color: string;
+  links: {
+    download: string;
+  };
+  urls: {
+    raw: string;
+    full: string;
+    small: string;
+    thumb: string;
+    regular: string;
+  };
+}
