+++ src/cards/LinkCard/types.ts
@@ -1,1 +1,3 @@
-export const CARD_TYPE = 'LINK';
+import { LINK_CARD_TYPE } from '../../api/database/types';
+
+export const CARD_TYPE = LINK_CARD_TYPE;
