+++ src/cards/NoteCard/types.ts
@@ -1,1 +1,3 @@
-export const CARD_TYPE = 'NOTE';
+import { NOTE_CARD_TYPE } from '../../api/database/types';
+
+export const CARD_TYPE = NOTE_CARD_TYPE;
